Simple UI components added; start/beginPlay for GameMap; loading assets from xml file

+++ G1/CollisionCalculation.cpp
@@ -21,7 +21,7 @@ namespace G1 {
 
 	CollisionCalculation::~CollisionCalculation()
 	{
-		for (int i = 0; i < collisionTypes.size(); i++) {
+		for (size_t i = 0; i < collisionTypes.size(); i++) {
 			delete collisionTypes[i];
 		}
 	}
